- Fixed MFString_CaseCmpN()
git-svn-id: https://fuji.googlecode.com/svn/trunk@305 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Source/MFHeap.cpp
@@ -161,14 +161,14 @@ void* operator new(size_t size)
 {
 //	MFDebug_Message(MFStr("new %d bytes", size));
 
-	return MFHeap_AllocInternal(size);
+	return MFHeap_AllocInternal((uint32)size);
 }
 
 void* operator new[](size_t size)
 {
 //	MFDebug_Message(MFStr("new %d bytes", size));
 
-	return MFHeap_AllocInternal(size);
+	return MFHeap_AllocInternal((uint32)size);
 }
 
 void operator delete(void *pMemory)
